+++ packages/components/src/Form/Fields/FieldSelectMulti/FieldSelectMulti.tsx
@@ -47,6 +47,7 @@ const FieldSelectMultiComponent = forwardRef(
         <SelectMulti
           {...omitFieldProps(props)}
           aria-describedby={`${id}-describedby`}
+          aria-labelledby={`${id}-labelledby`}
           id={id}
           validationType={validationMessage && validationMessage.type}
           ref={ref}