+++ packages/components/src/VisuallyHidden/VisuallyHidden.tsx
@@ -27,9 +27,9 @@
 import styled from 'styled-components'
 
 export const VisuallyHidden = styled.div`
-  position: absolute;
-  height: 1px;
-  width: 1px;
-  overflow: hidden;
   clip: rect(1px, 1px, 1px, 1px);
+  height: 1px;
+  overflow: hidden;
+  position: absolute;
+  width: 1px;
 `